// frontend/src/api/generated/types.gen.ts
export const RoundMode = {
    /**
     * ONE_VS_ONE
     */
    ONE_VS_ONE: 1,
    /**
     * TWO_VS_TWO
     */
    TWO_VS_TWO: 2,
    /**
     * THREE_VS_THREE
     */
    THREE_VS_THREE: 3,
    /**
     * FOUR_VS_FOUR
     */
    FOUR_VS_FOUR: 4,
    /**
     * FIVE_VS_FIVE
     */
    FIVE_VS_FIVE: 5,
    /**
     * SIX_VS_SIX
     */
    SIX_VS_SIX: 6
} as const;

/**
 * RoundMode
 */
export type RoundMode = typeof RoundMode[keyof typeof RoundMode];
// ...
export type SeasonRead = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Description
     */
    description: string;
    /**
     * Is Running
     */
    is_running: boolean;
    /**
     * Created At
     */
    created_at: string;
};